Feature/hide undo button

+++ www/js/controllers/remindersInboxCtrl.js
@@ -221,14 +221,17 @@ angular.module('starter')
 	    };
 
 		$scope.undo = function(){
+			$rootScope.showUndoButton = false;
 			var notificationsSyncQueue = localStorageService.getItemAsObject('notificationsSyncQueue');
+			if(!notificationsSyncQueue){
+				return false;
+			}
 			notificationsSyncQueue[0].trackingReminderNotification.hide = false;
 			localStorageService.addToOrReplaceElementOfItemByIdOrMoveToFront('trackingReminderNotifications',
 				notificationsSyncQueue[0].trackingReminderNotification);
 			localStorageService.deleteElementOfItemByProperty('notificationsSyncQueue',
 				'trackingReminderNotificationId', notificationsSyncQueue[0].trackingReminderNotificationId);
 			getTrackingReminderNotifications();
-			$rootScope.showUndoButton = false;
 		};
 
 		var getFilteredTrackingReminderNotifications = function(){
@@ -328,6 +331,9 @@ angular.module('starter')
 	    $scope.init = function(){
 			console.debug($state.current.name + ' initializing...');
 			$rootScope.stateParams = $stateParams;
+			if($rootScope.showUndoButton){
+				$rootScope.showUndoButton = false;
+			}
 			if (typeof Bugsnag !== "undefined") { Bugsnag.context = $state.current.name; }
 			if (typeof analytics !== 'undefined')  { analytics.trackView($state.current.name); }
 			showLoader();
